Lifted code from messagecard to app.
