+++ src/App.jsx
@@ -1,12 +1,44 @@
 import { MessageCard } from "./section/message-card/MessageCard"
-import { H1 } from "./styling/Typography"
+import {useState} from "react";
+import { CommentCard } from "./section/comment-card/CommentCard";
+import moment from 'moment';
 
 export const App = () => {
+  const [userInput, setUserInput] = useState("")
+  const [messages, setMessages] = useState([])
+
+  const submitHandler = (event) => {
+    event.preventDefault();
+
+    if (userInput.trim() === "") //The Trim takes away all the spaces
+      return
+
+    const newMessage = {
+      id: Date.now(), // This will be a uniqe ID
+      text: userInput.trim(),      
+      timestamp: moment().fromNow(),
+      likes: 0,
+    }
+
+    setMessages((prev) => [newMessage, ...prev])
+    setUserInput("")
+  }
+
+
   return (
     <>
-      <H1>Happy Thoughts</H1>
+      <h1 style={{ textAlign: 'center' }}>Happy Thoughts</h1>
       {/* <Header /> */}
-      < MessageCard />
+      < MessageCard input={userInput} setInput={setUserInput} comment={submitHandler}/>
+      {messages.map((message) => (
+        <>     
+           <CommentCard
+            key={message.id}
+            text={message.text}
+            timestamp={message.timestamp} /> 
+                   
+        </>
+      ))}
       {/* <Footer/> */}
     </>
   )
